make BSPServerService general
