+++ Sources/BSPServer/TaskManager/BSPTaskManager+BuildTargetCompile.swift
@@ -1,5 +1,5 @@
 //
-//  BSPTaskManager+ParsedBuild.swift
+//  BSPTaskManager+BuildTargetCompile.swift
 //
 //  Copyright © 2024 Wang Lun.
 //
@@ -7,10 +7,10 @@
 import BuildServerProtocol
 import Foundation
 
-extension BSPTaskManager {
+public extension BSPTaskManager {
     /// Execute build with progress parsed from xcodebuild output
     /// 这是最理想的方案 - 通过解析构建工具的输出来获得真实进度
-    public func executeBuildWithProgress(
+    func executeBuildWithProgress(
         using projectManager: any ProjectManager,
         targets: [BSPBuildTargetIdentifier],
         originId: String? = nil
@@ -58,23 +58,5 @@ extension BSPTaskManager {
             )
             throw error
         }
-    }
-
-    /// 解析 xcodebuild 输出来提取进度信息
-    private func parseXcodeBuildProgress(_ output: String) -> (progress: Double, message: String)? {
-        if output.contains("Building target") {
-            // 提取目标构建进度
-            return (0.3, "Building target...")
-        } else if output.contains("Compiling") {
-            // 提取文件编译进度
-            if output.range(of: #"\((\d+) of (\d+)\)"#, options: .regularExpression) != nil {
-                // TODO: 解析 "(5 of 20)" 格式并返回相应的进度百分比
-            }
-            return (0.6, "Compiling files...")
-        } else if output.contains("Linking") {
-            return (0.9, "Linking...")
-        }
-
-        return nil
     }
 }
